api_tasks/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
